refactor: split EscuelaContent into section components (#217)
