+++ src/features/escuela/index.ts
@@ -1,1 +1,9 @@
 export { EscuelaContent } from './components/EscuelaContent';
+export {
+  AutoridadesSection,
+  DocumentosSection,
+  EstadisticasSection,
+  HistoriaSection,
+  MisionVisionSection,
+  ValoresSection,
+} from './components/sections';
